feat: oAuth code optimized
